requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
mysql-connector-python